Development

This section introduces the source code that makes up the verifier.

Datatypes

The way to understand the software is to start by viewing the Datatypes module.

ElectionGuardVerifier.Datatypes — Module

Datatypes

Datatypes contains structs that reflect the contents of JSON files that make up election records.

Loader

The loader describes the expected directory structure of an election record.

ElectionGuardVerifier.Loader.load — Method

load(path::AbstractString)::Election_record

Load election records in the given directory.

Utililies

ElectionGuardVerifier.Utils — Module

Utils

Functions used to implement Election Guard checks.

ElectionGuardVerifier.Utils.mulpowmod — Method

mulpowermod(a, x, b, p) = (a * x ^ b) mod p

ElectionGuardVerifier.Utils.same — Method

same(c1::Constants, c2::Constants)::Bool

Are two sets of constants the same?

ElectionGuardVerifier.Utils.same — Method

same(c1::Ciphertext, c2::Ciphertext)::Bool

Are two ciphertexts the same?

ElectionGuardVerifier.Utils.within — Method

within(x::BigInt, p::BigInt)::Bool

Is 0 ≤ x < p?

ElectionGuardVerifier.Utils.within_mod — Function

within_mod(x::BigInt, q::BigInt, p::BigInt)::Bool

Is 0 ≤ x < p and (x ^ q) mod p == 1?

ElectionGuardVerifier.Utils.one_ct — Constant

one_ct::Ciphertext

one_ct = Ciphertext(1, 1)

ElectionGuardVerifier.Utils.prod_ct — Method

prod_ct(x1::Ciphertext, x2::Ciphertext, p::BigInt)::Ciphertext

Multiply two ciphertexts mod p

Check

The check function implements what is described in the version 1.0 ElectionGuard Specification.

ElectionGuardVerifier.check — Method

check(er::Election_record)::Bool

Check election records. Return true if all checks pass.

Errata

The places at which the code is at odds with what is in the spec can be found by searching for the follow comment in the Julia source code:

#! Spec conflict