Development

This section introduces the source code that makes up the verifier. According to C.A.R Hoare:

There are two ways of constructing a software design. One way is to make it so simple that there are obviously no deficiencies. And the other way is to make it so complicated that there are no obvious deficiencies.

Our goal is to always write code that is obviously correct.

Datatypes

The way to understand the software is to start by viewing the Datatypes module.

ElectionGuardVerifier.Datatypes — Module

Datatypes

This module contains the structures that appear in the JSON data files that make up an ElectionGuard election record. The field names in structs are the field names that appear in the JSON file with the exception of constants. The use of constants is too ubiquitous to use the verbose field names.

To see the structure of the data, it is best that you read the structs in this file in reverse order.

Loader

The loader describes the expected directory structure of an election record.

ElectionGuardVerifier.Loader — Module

Loader

The loader assumes the records are delivered using a standard directory structure that is encoded here. The structure is

manifest.json: Data about the election

constants.json: Constants used for ElGamal cryptography

context.json: Election parameters

coefficients.json: Lagrange coefficients

guardians/*.json: Per guardian data

encryption_devices/*.json: Encryption devices

submitted_ballots/*.json: Submitted ballots

spoiled_ballots/*.json: Spoiled ballots (may be missing)

encrypted_tally: Encrypted tally

tally.json: The tally

ElectionGuardVerifier.Loader.load — Method

load(path::AbstractString)::Election_record

Load election records in the given directory.

Utililies

ElectionGuardVerifier.Utils — Module

Utils

Functions used to implement Election Guard checks.

ElectionGuardVerifier.Utils.mulpowmod — Method

mulpowermod(a, x, b, p) = (a * x ^ b) mod p

ElectionGuardVerifier.Utils.same — Method

same(c1::Constants, c2::Constants)::Bool

Are two sets of constants the same?

ElectionGuardVerifier.Utils.same — Method

same(c1::Ciphertext, c2::Ciphertext)::Bool

Are two ciphertexts the same?

ElectionGuardVerifier.Utils.within — Method

within(x::BigInt, p::BigInt)::Bool

Is 0 ≤ x < p?

ElectionGuardVerifier.Utils.within_mod — Function

within_mod(x::BigInt, q::BigInt, p::BigInt)::Bool

Is 0 ≤ x < p and (x ^ q) mod p == 1?

ElectionGuardVerifier.Utils.one_ct — Constant

one_ct::Ciphertext

one_ct = Ciphertext(1, 1)

ElectionGuardVerifier.Utils.prod_ct — Method

prod_ct(x1::Ciphertext, x2::Ciphertext, p::BigInt)::Ciphertext

Multiply two ciphertexts mod p

Answers

ElectionGuardVerifier.Answers — Module

Answers

This module provides data structures and operations used to report verification answers. At the top-level, there is a verification record. It identifies the specification version, the election, and contains a list of verification answers. A verification answer is the result of checking all or part of a verification step as defined in an ElectionGuard specification.

A verification answer contains a verification step number, a string listing items that failed while verifying the step, the step title, a comment, the number of records checked, and the number of records that failed. A failing step that has no items is marked using item "X".

In the verification routies, each verification item check returns an integer. The integer is zero if the check passes, otherwise it is an integer with one bit set. The bit is used to identify the item being checked. The bit patterns are exported as constants A, B, C, D, E, F, G, H, I, J, and K. The bit patterns for multiple checks are combined using bitwise or.

ElectionGuardVerifier.Answers.answer — Method

Construct a step answer

ElectionGuardVerifier.Answers.verification_record — Method

Construct a verification record

ElectionGuardVerifier.Answers.bits2items — Method

Convert a bit pattern to items

Check

The check function implements what is described in the version 1.0 ElectionGuard Specification.

ElectionGuardVerifier.check — Method

check(er::Election_record, path::String="")::Bool

Check election record. Write answers to path in JSON if path is not empty.